BackEnd/requirements.txt
boto3
fastapi
html2text
ipython
llama_index
llama-index-vector-stores-qdrant
fastembed
markdown2
openai
pandas
pdfkit
Pillow
pydantic
python-dotenv
PyYAML
qdrant_client
Requests
uvicorn
torch
transformers